share unsafe UUID builder logic between js and jvm
